only add to chain if building on tip
if not building on tip we shouldn't add to chain.

Situation would happen if we are reindexing blocks and we are < maxtipage*2 so we won't hear about nevm block data (which means we only add to mappings). In this case we will end up skipping over blocks and when it comes to >= maxtipage*2 and we get nevm block data we will not be in sequential order. Instead we rely on Geth sync to fetch blocks and we only setup mappings (not insert into chain even if we have the block data)

+++ eth/backend.go
@@ -26,7 +26,6 @@ import (
 	"sync"
 	"sync/atomic"
 	"time"
-	"encoding/hex"
 
 	"github.com/ethereum/go-ethereum/accounts"
 	"github.com/ethereum/go-ethereum/common"
@@ -392,10 +391,16 @@ func New(stack *node.Node, config *ethconfig.Config) (*Ethereum, error) {
 		// add before potentially inserting into chain (verifyHeader depends on the mapping), we will delete if anything is wrong
 		eth.blockchain.WriteNEVMMappings(nevmBlockConnect.Sysblockhash, nevmBlockConnect.Blockhash)
 		if nevmBlockConnect.Block != nil {
-			_, err := eth.blockchain.InsertChain(types.Blocks([]*types.Block{nevmBlockConnect.Block}))
-			if err != nil {
-				eth.blockchain.DeleteNEVMMappings(nevmBlockConnect.Sysblockhash, nevmBlockConnect.Blockhash)
-				return err
+			// insert into chain if building on the tip, otherwise just add into mapping and fetch via normal sync via geth
+			current := eth.blockchain.CurrentBlock()
+			if current.Hash() == nevmBlockConnect.Block.ParentHash() {
+				_, err := eth.blockchain.InsertChain(types.Blocks([]*types.Block{nevmBlockConnect.Block}))
+				if err != nil {
+					eth.blockchain.DeleteNEVMMappings(nevmBlockConnect.Sysblockhash, nevmBlockConnect.Blockhash)
+					return err
+				}
+			} else {
+				log.Info("not building on tip, add to mapping...", "blocknumber", nevmBlockConnect.Block.NumberU64(), "currenthash", current.Hash().String(), "proposedparenthash", nevmBlockConnect.Block.ParentHash().String())
 			}
 		}
 		return nil
